Rename src directory to nanodbc.
Switch to buried header and #include <nanodbc/nanodbc.h>
Rename directories with singular names.
Rename scripts to utility.
