+++ .gitignore
@@ -20,4 +20,4 @@
 /docker_build/
 example
 example.dSYM/
-/examples/empty.cpp
+/example/empty.cpp
