+++ .travis.yml
@@ -141,11 +141,11 @@ before_install:
       echo "$TRAVIS_COMMIT_MESSAGE" | grep -E  '\[(skip travis|travis skip)\]'
       && echo "[skip travis] has been found, exiting."
       && exit 0 || true
-  - ${TRAVIS_BUILD_DIR}/scripts/ci/travis/before_install.$TRAVIS_OS_NAME.sh
-  - ${TRAVIS_BUILD_DIR}/scripts/ci/travis/before_install.vertica.sh
+  - ${TRAVIS_BUILD_DIR}/utility/ci/travis/before_install.$TRAVIS_OS_NAME.sh
+  - ${TRAVIS_BUILD_DIR}/utility/ci/travis/before_install.vertica.sh
 
 before_script:
-  - source "${TRAVIS_BUILD_DIR}/scripts/ci/travis/before_script.$DB.sh"
+  - source "${TRAVIS_BUILD_DIR}/utility/ci/travis/before_script.$DB.sh"
   - if [[ -z ${BUILD_SHARED_LIBS+x} ]]; then export BUILD_SHARED_LIBS=OFF; fi
   - if [[ -f /etc/odbcinst.ini ]]; then export ODBCSYSINI=/etc; fi
   - if [[ -f /etc/odbc.ini ]]; then export ODBCINI=/etc/odbc.ini; fi
@@ -154,4 +154,4 @@ before_script:
   - if [[ -s "$ODBCINI" ]]; then odbcinst -q -s; fi
 
 script:
-  - ${TRAVIS_BUILD_DIR}/scripts/ci/travis/script.sh
+  - ${TRAVIS_BUILD_DIR}/utility/ci/travis/script.sh
