+++ utility/ci/travis/before_script.sqlite.sh
